+++ src/main/java/org/javlo/config/StaticConfig.java
@@ -1905,7 +1905,7 @@ public class StaticConfig extends Observable {
 	}
 
 	public String getImageFormat() {
-		return properties.getString("content.image-format", "png,jpg,jpeg,gif,svg");
+		return properties.getString("content.image-format", "png,jpg,jpeg,gif,svg,webp");
 	}
 
 	public String getVideoFormat() {